Fixes #1911 (move reading plan import to selector)
Move the existing ZIP import action to the reading plan selection screen without changing the import format or behavior.
