+++ app/src/main/res/menu/reading_plan_selector.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!--
-  ~ Copyright (c) 2022-2022 Martin Denham, Tuomas Airaksinen and the AndBible contributors.
+  ~ Copyright (c) 2026 Martin Denham, Tuomas Airaksinen and the AndBible contributors.
   ~
   ~ This file is part of AndBible: Bible Study (http://github.com/AndBible/and-bible).
   ~
@@ -16,18 +16,8 @@
   ~ If not, see http://www.gnu.org/licenses/.
   -->
 
-<!-- The main menu accessed via Menu key
- 
- @author Martin Denham [mjdenham at gmail dot com]
--->
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <group android:id="@+id/group1">
-		<item android:id="@+id/setCurrentDay"
-			  android:title="@string/set_current_day"/>
-		<item android:id="@+id/setStartDate"
-			  android:title="@string/rdg_plan_set_start_date"/>
-		<item android:id="@+id/reset"
-			  android:icon="@drawable/ic_baseline_undo_24"
-			  android:title="@string/reset_generic"/>
-	</group>
+    <item
+        android:id="@+id/import_reading_plan"
+        android:title="@string/import_reading_plan" />
 </menu>
